Move toolbar code from shell page to toolbar view model
