+++ Files/IShellPage.cs
@@ -36,6 +36,16 @@ namespace Files
         void UpdatePathUIToWorkingDirectory(string newWorkingDir, string singleItemOverride = null);
 
         void NavigateToPath(string navigationPath, Type sourcePageType, NavigationArguments navArgs = null);
+        /// <summary>
+        /// Gets the layout mode for the specified path then navigates to it
+        /// </summary>
+        /// <param name="navigationPath"></param>
+        /// <param name="navArgs"></param>
+        public void NavigateToPath(string navigationPath, NavigationArguments navArgs = null);
+        /// <summary>
+        /// Navigates to the home page
+        /// </summary>
+        public void NavigateHome();
 
         void NavigateWithArguments(Type sourcePageType, NavigationArguments navArgs);
